// src/types.ts
type IdRequest = string | string
export type PersonUserObjectResponse = {
  type: 'person'
  person: {
    email?: string
  }
  name: string | null
  avatar_url: string | null
  id: IdRequest
  object: 'user'
}
type EmptyObject = Record<string, never>
export type PartialUserObjectResponse = {
  id: IdRequest
  object: 'user'
}
// ...
type TextRequest = string
// ...
export type TextRichTextItemResponse = {
  type: 'text'
  text: {
    content: string
    link: {
      url: TextRequest
    } | null
  }
  annotations: AnnotationResponse
  plain_text: string
  href: string | null
}
// ...
export type RichTextItemResponse = TextRichTextItemResponse | MentionRichTextItemResponse | EquationRichTextItemResponse
// ...
export type PartialPageObjectResponse = {
  object: 'page'
  id: string
}
// ...
export type PartialDatabaseObjectResponse = {
  object: 'database'
  id: string
  properties: Record<string, DatabasePropertyConfigResponse>
}
export type DatabaseObjectResponse = {
  title: Array<RichTextItemResponse>
  description: Array<RichTextItemResponse>
  icon:
    | {
        type: 'emoji'
        emoji: EmojiRequest
      }
    | null
    | {
        type: 'external'
        external: {
          url: TextRequest
        }
      }
    | null
    | {
        type: 'file'
        file: {
          url: string
          expiry_time: string
        }
      }
    | null
  cover:
    | {
        type: 'external'
        external: {
          url: TextRequest
        }
      }
    | null
    | {
        type: 'file'
        file: {
          url: string
          expiry_time: string
        }
      }
    | null
  properties: Record<string, DatabasePropertyConfigResponse>
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  created_by: PartialUserObjectResponse
  last_edited_by: PartialUserObjectResponse
  is_inline: boolean
  object: 'database'
  id: string
  created_time: string
  last_edited_time: string
  archived: boolean
  url: string
  public_url: string | null
}
export type PartialBlockObjectResponse = {
  object: 'block'
  id: string
}
type ApiColor =
  | 'default'
  | 'gray'
  | 'brown'
  | 'orange'
  | 'yellow'
  | 'green'
  | 'blue'
  | 'purple'
  | 'pink'
  | 'red'
  | 'gray_background'
  | 'brown_background'
  | 'orange_background'
  | 'yellow_background'
  | 'green_background'
  | 'blue_background'
  | 'purple_background'
  | 'pink_background'
  | 'red_background'
export type ParagraphBlockObjectResponse = {
  type: 'paragraph'
  paragraph: {
    rich_text: Array<RichTextItemResponse>
    color: ApiColor
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
export type Heading1BlockObjectResponse = {
  type: 'heading_1'
  heading_1: {
    rich_text: Array<RichTextItemResponse>
    color: ApiColor
    is_toggleable: boolean
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
export type Heading2BlockObjectResponse = {
  type: 'heading_2'
  heading_2: {
    rich_text: Array<RichTextItemResponse>
    color: ApiColor
    is_toggleable: boolean
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
export type Heading3BlockObjectResponse = {
  type: 'heading_3'
  heading_3: {
    rich_text: Array<RichTextItemResponse>
    color: ApiColor
    is_toggleable: boolean
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
export type BulletedListItemBlockObjectResponse = {
  type: 'bulleted_list_item'
  bulleted_list_item: {
    rich_text: Array<RichTextItemResponse>
    color: ApiColor
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
export type NumberedListItemBlockObjectResponse = {
  type: 'numbered_list_item'
  numbered_list_item: {
    rich_text: Array<RichTextItemResponse>
    color: ApiColor
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
export type QuoteBlockObjectResponse = {
  type: 'quote'
  quote: {
    rich_text: Array<RichTextItemResponse>
    color: ApiColor
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
export type ToDoBlockObjectResponse = {
  type: 'to_do'
  to_do: {
    rich_text: Array<RichTextItemResponse>
    color: ApiColor
    checked: boolean
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
export type ToggleBlockObjectResponse = {
  type: 'toggle'
  toggle: {
    rich_text: Array<RichTextItemResponse>
    color: ApiColor
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
export type TemplateBlockObjectResponse = {
  type: 'template'
  template: {
    rich_text: Array<RichTextItemResponse>
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
// ...
export type ChildPageBlockObjectResponse = {
  type: 'child_page'
  child_page: {
    title: string
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
// ...
type LanguageRequest =
  | 'abap'
  | 'agda'
  | 'arduino'
  | 'assembly'
  | 'bash'
  | 'basic'
  | 'bnf'
  | 'c'
  | 'c#'
  | 'c++'
  | 'clojure'
  | 'coffeescript'
  | 'coq'
  | 'css'
  | 'dart'
  | 'dhall'
  | 'diff'
  | 'docker'
  | 'ebnf'
  | 'elixir'
  | 'elm'
  | 'erlang'
  | 'f#'
  | 'flow'
  | 'fortran'
  | 'gherkin'
  | 'glsl'
  | 'go'
  | 'graphql'
  | 'groovy'
  | 'haskell'
  | 'html'
  | 'idris'
  | 'java'
  | 'javascript'
  | 'json'
  | 'julia'
  | 'kotlin'
  | 'latex'
  | 'less'
  | 'lisp'
  | 'livescript'
  | 'llvm ir'
  | 'lua'
  | 'makefile'
  | 'markdown'
  | 'markup'
  | 'matlab'
  | 'mathematica'
  | 'mermaid'
  | 'nix'
  | 'notion formula'
  | 'objective-c'
  | 'ocaml'
  | 'pascal'
  | 'perl'
  | 'php'
  | 'plain text'
  | 'powershell'
  | 'prolog'
  | 'protobuf'
  | 'purescript'
  | 'python'
  | 'r'
  | 'racket'
  | 'reason'
  | 'ruby'
  | 'rust'
  | 'sass'
  | 'scala'
  | 'scheme'
  | 'scss'
  | 'shell'
  | 'solidity'
  | 'sql'
  | 'swift'
  | 'toml'
  | 'typescript'
  | 'vb.net'
  | 'verilog'
  | 'vhdl'
  | 'visual basic'
  | 'webassembly'
  | 'xml'
  | 'yaml'
  | 'java/c/c++/c#'
export type CodeBlockObjectResponse = {
  type: 'code'
  code: {
    rich_text: Array<RichTextItemResponse>
    caption: Array<RichTextItemResponse>
    language: LanguageRequest
  }
  parent:
    | {
        type: 'database_id'
        database_id: string
      }
    | {
        type: 'page_id'
        page_id: string
      }
    | {
        type: 'block_id'
        block_id: string
      }
    | {
        type: 'workspace'
        workspace: true
      }
  object: 'block'
  id: string
  created_time: string
  created_by: PartialUserObjectResponse
  last_edited_time: string
  last_edited_by: PartialUserObjectResponse
  has_children: boolean
  archived: boolean
}
// ...
export type BlockObjectResponse =
  | ParagraphBlockObjectResponse
  | Heading1BlockObjectResponse
  | Heading2BlockObjectResponse
  | Heading3BlockObjectResponse
  | BulletedListItemBlockObjectResponse
  | NumberedListItemBlockObjectResponse
  | QuoteBlockObjectResponse
  | ToDoBlockObjectResponse
  | ToggleBlockObjectResponse
  | TemplateBlockObjectResponse
  | SyncedBlockBlockObjectResponse
  | ChildPageBlockObjectResponse
  | ChildDatabaseBlockObjectResponse
  | EquationBlockObjectResponse
  | CodeBlockObjectResponse
  | CalloutBlockObjectResponse
  | DividerBlockObjectResponse
  | BreadcrumbBlockObjectResponse
  | TableOfContentsBlockObjectResponse
  | ColumnListBlockObjectResponse
  | ColumnBlockObjectResponse
  | LinkToPageBlockObjectResponse
  | TableBlockObjectResponse
  | TableRowBlockObjectResponse
  | EmbedBlockObjectResponse
  | BookmarkBlockObjectResponse
  | ImageBlockObjectResponse
  | VideoBlockObjectResponse
  | PdfBlockObjectResponse
  | FileBlockObjectResponse
  | AudioBlockObjectResponse
  | LinkPreviewBlockObjectResponse
  | UnsupportedBlockObjectResponse